chore(common): Move outputs.groundwork slog adapter to common
